% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_processing.R
\name{akima_bilinear}
\alias{akima_bilinear}
\title{Bilinear data.frame interpolation}
\usage{
akima_bilinear(df, out_dim)
}
\arguments{
\item{df}{data.frame containing gridded x, y locations and z values}

\item{out_dim}{vector containing \code{c(nx, ny)} as number of x and y output locations}
}
\description{
Interpolated on a grid using akima interpolation
}
\details{
The data.frame df must contain columns for x, y, and z. The x and y
values must be on a grid.

Interpolation is done using \code{akima::bilinear}
}
\examples{
library(ggplot2)
library(ggp)
df <- expand.grid(x=1:5,y=1:5)
df$z <- df$x^2 + df$y^2
df_out <- akima_bilinear(df, c(100,100))
ggplot(df) + geom_raster(aes(x,y,fill=z)) + coord_equal()
ggplot(df_out) + geom_raster(aes(x,y,fill=z)) + coord_equal()
}
