% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{bound_vec}
\alias{bound_vec}
\title{Bound a numeric vector}
\usage{
bound_vec(x, bounds)
}
\arguments{
\item{x}{A number or numeric vector}

\item{bounds}{A vector \code{c(a,b)} which identify the lower (a) and upper (b) limits, respectively.}
}
\description{
Bound a numeric vector with upper and lower limits, so that anything outside the limit is replaced by the limit.
}
\examples{
bound_vec(1:20,c(2.5,10.5))
dplyr::mutate(data.frame(a=1:20),b=bound_vec(a,c(3.2,15.7)))
}
