% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{crore_to_usd}
\alias{crore_to_usd}
\title{Convert Crore INR (Rs) to USD (\$)}
\usage{
crore_to_usd(crore, rs_per_usd = 70)
}
\arguments{
\item{crore}{Amount of rupees in crore}

\item{rs_per_usd}{Conversion rate: rupees per dollar}
}
\description{
Convert crore to USD. 1 crore = 1 x 10^7
}
\examples{
crore_to_usd(70,70) # 70 crore Rs is $10M (1e7) with exchange rate of 70 Rs = 1 USD
}
