% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{fig_get_gdrive_path}
\alias{fig_get_gdrive_path}
\title{Get Google Drive output path}
\usage{
fig_get_gdrive_path(fig_path = NULL)
}
\arguments{
\item{fig_path}{Path to figures directory -- if NULL, function will search repo for it}
}
\description{
Search for Google Drive symlink in results/figure, and return relative symlink path
}
\examples{
\dontrun{
output_dir_path <- fig_get_gdrive_path()
output_dir_path <- fig_get_gdrive_path(fig_path)
}
}
