% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{fig_set_output}
\alias{fig_set_output}
\title{Set file output path}
\usage{
fig_set_output(group_name, fig_path = NULL)
}
\description{
Set, create, and output the Google drive output path
}
\examples{
output_dir_path <- set_fig_output("my_script_name")
}
