% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_plotting.R
\name{gg_align_lr}
\alias{gg_align_lr}
\title{GG align left and right}
\usage{
gg_align_lr(p1, p2)
}
\value{
from here: https://stackoverflow.com/questions/17736434/aligning-distinct-non-facet-plots-in-ggplot2-using-rpy2-in-python/17768224#17768224

from here: https://stackoverflow.com/questions/17736434/aligning-distinct-non-facet-plots-in-ggplot2-using-rpy2-in-python/17768224#17768224
}
\description{
GG align left and right
}
