% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing.R
\name{gis_longitude_to_utm_zone}
\alias{gis_longitude_to_utm_zone}
\title{Get UTM zone from coordinates}
\usage{
gis_longitude_to_utm_zone(lon)
}
\arguments{
\item{lon}{longitude in [-180,180]}
}
\value{
The number of the UTM zone containing the longitude
}
\description{
Get UTM zone from coordinates
}
\examples{
gis_longitude_to_utm_zone(-45)
}
