% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing.R
\name{gis_utm_zone_to_proj4}
\alias{gis_utm_zone_to_proj4}
\title{Convert utm zone to proj4string}
\usage{
gis_utm_zone_to_proj4(utm_zone)
}
\arguments{
\item{utm_zone}{UTM zone as numeric integer, or string}
}
\value{
The proj4string of the UTM Zone
}
\description{
Convert utm zone to proj4string
}
\examples{
gis_utm_zone_to_proj4(32)
gis_utm_zone_to_proj4(gis_longitude_to_utm_zone(-45))
}
