% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_plotting.R
\name{\%=\%}
\alias{\%=\%}
\title{Set equal to list}
\usage{
l \%=\% r
}
\description{
Set equal to list. from: http://stackoverflow.com/questions/7519790/assign-multiple-new-variables-in-a-single-line-in-r
}
