% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_plotting.R
\docType{data}
\name{my_colors}
\alias{my_colors}
\title{Get colors from colorspace
Get set of colors for plotting
param name The name that matches a corresponding color group. Must be
one of the following:}
\format{An object of class \code{list} of length 4.}
\usage{
my_colors
}
\description{
Get colors from colorspace
Get set of colors for plotting
param name The name that matches a corresponding color group. Must be
one of the following:
}
\keyword{datasets}
