% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_setup.R
\name{repo_init}
\alias{repo_init}
\title{Create a repository with GP standard structure}
\usage{
repo_init(path)
}
\description{
Create a project repository with GP standard structure. Should be done INSIDE
the dirctory that will house the repository.
}
\examples{
setwd("path/to/repo")
repo_init()
}
