% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_mapping.R
\name{scalebar_geoms}
\alias{scalebar_geoms}
\title{Add scalebar to a gglpot}
\usage{
scalebar_geoms(x_start, x_diff_m, y_start, y_diff, p4s_orig, p4s_utm,
  slabs)
}
\description{
Add scalebar to a gglpot // deprecated. Use ggspatial
}
