% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{fig_get_gdrive_path}
\alias{fig_get_gdrive_path}
\title{Get Google Drive output path}
\usage{
fig_get_gdrive_path(fig_path = NULL)
}
\description{
Search for Google Drive symlink in results/figure, and return relative symlink path
}
\examples{
output_dir_path <- get_gdrive_fig_path()
output_dir_path <- get_gdrive_fig_path(fig_path)
}
