% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{print_data_frame_for_entry}
\alias{print_data_frame_for_entry}
\title{Print data frame for re-entry}
\usage{
print_data_frame_for_entry(df, strings_as_factors = FALSE,
  single_line = FALSE)
}
\description{
Print data frame for re-entry
}
\examples{
df <- data.frame(a=1:4,b=11:14,c=c("a","b","c","q"))
ggp::print_data_frame_for_entry(df)
ggp::print_data_frame_for_entry(df,single_line=T)
df <- data.frame(a=1,b=14,c=c("a"))
ggp::print_data_frame_for_entry(df)
bounds <- tibble::tibble(bID=c(5, 6, 7, 8),
                    x1=c(468888, 572670, 483978, 468888),
                    x2=c(572670, 588746, 588746, 483978),
                    y1=c(4592114, 4630407, 4518117, 4592114),
                    y2=c(4630407, 4556624, 4556624, 4518117),
                    v=c("foo","bar","oof","rab"))
bounds \%>\% ggp::print_data_frame_for_entry()
bounds \%>\% ggp::print_data_frame_for_entry(single_line=TRUE)
df <- data.frame(bID=c(c(5, 6, 7, 8)),
                 x1=c(c(-87.44, -86.18, -85.88, -87.13)),
                 y1=c(c(41.46, 41.81, 41.17, 40.83)),
                 x2=c(c(-86.18, -85.88, -87.13, -87.44)),
                 y2=c(c(41.81, 41.17, 40.83, 41.46)),
                 bound_type=c(c(1, 1, 1, 1)))
df \%>\% print_data_frame_for_entry()
}
