% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_setup.R
\name{repo_create_gdrive_symlink}
\alias{repo_create_gdrive_symlink}
\title{Create symbolic link to gdrive}
\usage{
repo_create_gdrive_symlink(link_path = NULL, gdrive_path)
}
\description{
Create symbolic link to gdrive
}
\examples{
repo_create_gdrive_symlink(gdrive_path="~/Google Drive/my_project/fig")
}
